add cleanup + uploading features (#99)
This commit allows the user to specify cleanup and upload command line
options. The cleanup option allows the user to run the tests with space
efficiency (removing mlir and vmfb files by default). Which types of
files are removed can be changed by the user. Also, there is an
uploading feature introduced. This allows the user to specify a list of
tests in the `gold/upload_list.txt` file. If the test that the user is
running falls in this list, the file types that the user choses are
uploaded to an Azure space. They are organized here with unique
identifier based on test name + date + UTC time so we won't try
uploading to the same name. Also, after running the suite, a
`upload_urls.txt` file is created. In this file, you can find a mapping
of the test name to Azure link. This way the user doesn't have to deal
with Azure. More details can be found in code.