add output selection and postprocessing for more accurate validation (#108)
This commit addresses mismatches we were seeing at inference stage.
There were two main problems. We were comparing every output generated
by the model, when we should just be looking at the important outputs
(ones that influence our prediction). Based on whether there is a head
attached to the bare model, this is either the logits or last hidden
state. So added support so we can compare a specific output (modifying
the e2eshark dict and run.py). Next problem was with the logits. We
should be comparing the normalized logits for a model output. This
required adding a SoftMax postprocessing pass (normalizing the `(B, S,
V)` logit output over the `V` dimension so we have a probability
distribution of all possible tokens at all sequence positions). I didn't
want to add topk to the postprocessing because that would hardcode us to
specific model configuration/prompts. It might work for a certain
configuration, but the inaccuracies in prediction we ignored can lead to
inconsistencies with other configurations. Lastly, I added support so
that user is notified with a message when there are NaN values in the
output tensors, so the user is aware of the exact numeric issues.