Support for sdxl pipeline (benchmarking) (#155)
This commit adds the V0 support for testing all sdxl submodels. This PR
is not a long term solution for benchmarking as Scott and I discussed
here: https://github.com/nod-ai/SHARK-TestSuite/pull/152. This is the
result of a request to get sdxl benchmarking in ASAP by our team. Due to
the high priority for this to be added to the sdxl testing as our team
lands patches in IREE, this is simply just to get the implementation in
and working. Scott and I discussed some more intensive and well
structured ways to add benchmarking, which either of us may implement in
the future.

Also, this PR depends on this one in terms of landing:
https://github.com/nod-ai/SHARK-TestSuite/pull/152 (hence the CI
failure)

Notes for future if we decide that we need a stronger implementation: 

1. Maybe something like https://github.com/openxla/iree/pull/16965 which
will feed into https://perf.iree.dev/.
2. This is the benchmarking framework we already have:
https://github.com/openxla/iree-comparative-benchmark and
https://github.com/openxla/iree/tree/main/build_tools/benchmarks
3. Some questions Scott raised to keep in mind for future
implementation:
* What metrics/artifacts do we want from benchmarking?
  * Each model in isolation? Full pipeline latency? Just dispatch time?
* What do we want done with benchmark results / artifacts?
* The in-tree benchmarks in IREE submit results to a dashboard (that
should use a queryable database...), upload Tracy files to cloud
storage, and comment on pending pull requests with results summaries
* Where do we want benchmarks to run?
  * Right after tests, on presubmit to IREE?
  * In a separate job, on separate runners?
If we decide benchmarking needs changes, we will address all of these
and come up with a more structured, methodical implementation that
either creates a new benchmarking flow here or plugs into the iree
benchmarking setup.