Add a design alternative for e2eshark (#214)
The broad goal of this design rework is to make it easier to add
functionality and new tests to e2eshark. Towards this, there are
sub-goals:

1. break up the complexity of run.py into a more coherent file
structure.
2. remove the need to generate excessive boilerplate for model.py files,
and generally aim to make adding tests as simple as possible.
3. avoid running command-line scripts through python. This includes
transitioning to using python bindings for torch-mlir and iree, and
developing an import-friendly alternative to model.py files.
4. avoid concatenating python files (again, by building a better import
structure).
5. minimize redundancy in the log (test-run) directories. For example, I
believe the input and output tensors are actually stored three different
ways (as a .bin, .pt, and .pkl), and this should not need to be the
case.

Abstractly there seem to be four input pieces to manage:

1. The model
2. The choice of frontend importer (e.g. iree-import-onnx, torch_mlir's
fx importer, etc)
3. the choice of backend (e.g. compiling with iree from torch)
4. the input

Individual tests should just need to determine how to access or generate
the model and how to generate a sample input.
Command line arguments should determine which frontend/backend
combination is to be used during the test.

The desired output pieces seem to be:

1. input/output/golden_output tensors
2. mlir/vmfb files used to generate output
3. stage and config-specific debugging info for tests which fail

### Current high-priority TODOs:

1. look into using pytest or another testing framework.
2. generate command-line reproducers for ease of debugging individual
stages
3. come up with a better name for the directory than "alt_e2eshark".
4. identify functionality present in e2eshark that is currently lacking
and should be migrated over. E.g. Chi's comment about the --report flag.
5. add more type hints to make it easier to track down definitions.
6. add (much) more to the README about adding tests and running them
with various command line options.
7. look into removing torch-mlir as a python dependency entirely
(perhaps optionally, so torch-mlir developers can still use the test
suite for debugging op conversions).
8. Consider alternative options for storing cache directory paths, log
directory paths, onnx model paths, etc. OnnxModelInfo is currently a bit
clunky in needing to sometimes infer the log directory from the model
path, etc.